fix: 收费项单价可以小数

+++ yudao-module-jl/yudao-module-jl-biz/src/main/java/cn/iocoder/yudao/module/jl/entity/project/ProjectChargeitem.java
@@ -87,7 +87,7 @@ public class ProjectChargeitem extends BaseEntity {
      * 单价
      */
     @Column(name = "unit_fee", nullable = false )
-    private Integer unitFee;
+    private BigDecimal unitFee;
 
     @Column(name = "spec", nullable = false )
     private String spec;
